Renamed client.h/cpp to context.h/cpp
