+++ libraries/vircadia-client/src/context.cpp
@@ -9,7 +9,7 @@
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 
-#include "client.h"
+#include "context.h"
 
 #include <nlohmann/json.hpp>
